update install and uninstall scripts

+++ install.sh
@@ -6,8 +6,9 @@ if [ $EUID -ne 0 ]; then
    exit 1
 fi
 
-mkdir -p /etc/localrepo
-cp packages.conf /etc/localrepo
+mkdir -p /etc/localrepo/scripts
+cp geturl-plex /etc/localrepo/scripts
+[ ! -f /etc/localrepo/packages.conf ] && cp packages.conf /etc/localrepo
 
 cp update-localrepo /usr/local/bin
 chmod a+x /usr/local/bin/update-localrepo
@@ -18,6 +19,9 @@ cp update-localrepo.service update-localrepo.timer /var/lib/localrepo
 systemctl link /var/lib/localrepo/update-localrepo.timer /var/lib/localrepo/update-localrepo.service
 systemctl enable --now update-localrepo.timer
 
+systemctl start update-localrepo.service
+
 echo -e "Installation done.\n"
-echo "To start the first repo update please run:"
-echo "sudo systemctl start update-localrepo.service"
+echo "You can edit the package list in /etc/localrepo/packages.conf"
+echo "To manually update the local repository, run:"
+echo "  sudo update-localrepo"
